requirements.txt
accelerate
datasets
evaluate
fastcore
openai
transformers
python-dotenv
pandas